MAINT: Move ignore_nanwarnings to numpy_utils.
